#121 Pacotes como interfaces, dao e repositorio movidos de model para petpalacy
