+++ src/main/java/br/edu/ifpe/petpalacy/interfaces/InterfaceEmpresa.java
@@ -21,7 +21,7 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  SOFTWARE.
  */
-package br.edu.ifpe.petpalacy.model.interfaces;
+package br.edu.ifpe.petpalacy.interfaces;
 
 /**
  *
