+++ src/main/java/br/edu/ifpe/petpalacy/interfaces/InterfaceGenerico.java
@@ -21,15 +21,19 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  SOFTWARE.
  */
-package br.edu.ifpe.petpalacy.model.interfaces;
+package br.edu.ifpe.petpalacy.interfaces;
+
+import java.util.List;
 
 /**
  *
  * @author Kaio César Bezerra da Silva <kaio_gus@outlook.com>
- * @param <Cliente>
+ * @param <E>
  */
-public interface InterfaceCliente<Cliente> extends InterfaceGenerico<Cliente> {
-    
-    public Cliente autenticar(String login, String senha);
-    public Cliente buscarCpf(String cpf);
+public interface InterfaceGenerico<E> {
+    public void salvar(E e) throws Exception;
+    public E buscar(Integer codigo);
+    public void editar(E e) throws Exception;
+    public void deletar(E e) throws Exception;
+    public List<E> listar();
 }
