+++ src/main/java/br/edu/ifpe/petpalacy/repositorio/implementacao/RepositorioEmpresaImplDB.java
@@ -21,11 +21,11 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  SOFTWARE.
  */
-package br.edu.ifpe.petpalacy.model.repositorio.implementacao;
+package br.edu.ifpe.petpalacy.repositorio.implementacao;
 
-import br.edu.ifpe.petpalacy.model.dao.PersistenciaDAO;
+import br.edu.ifpe.petpalacy.dao.PersistenciaDAO;
 import br.edu.ifpe.petpalacy.model.entidades.Empresa;
-import br.edu.ifpe.petpalacy.model.interfaces.InterfaceEmpresa;
+import br.edu.ifpe.petpalacy.interfaces.InterfaceEmpresa;
 import java.util.List;
 
 /**
