+++ src/test/java/br/edu/ifpe/petpalacy/model/repositorio/implementacao/RepositorioServicoImplDBTest.java
@@ -23,8 +23,8 @@ package br.edu.ifpe.petpalacy.model.repositorio.implementacao;
 import br.edu.ifpe.petpalacy.model.entidades.Empresa;
 import br.edu.ifpe.petpalacy.model.entidades.Endereco;
 import br.edu.ifpe.petpalacy.model.entidades.Servico;
-import br.edu.ifpe.petpalacy.model.repositorio.implementacao.RepositorioServicoImplDB;
-import br.edu.ifpe.petpalacy.model.repositorio.implementacao.RepositorioServicoImplDB;
+import br.edu.ifpe.petpalacy.repositorio.implementacao.RepositorioServicoImplDB;
+import br.edu.ifpe.petpalacy.repositorio.implementacao.RepositorioServicoImplDB;
 import br.edu.ifpe.petpalacy.util.HibernateUtil;
 import java.math.BigDecimal;
 import java.util.ArrayList;
